ai/requirements.txt
# VLM 파인튜닝 및 데이터 생성용 의존성
# pip install -r requirements.txt

# 데이터 생성
faker>=18.0.0
tqdm>=4.65.0
requests>=2.28.0

# 모델 학습 (LLaMA-Factory)
transformers==4.57.1
huggingface_hub<1.0.0
accelerate>=1.3.0
peft>=0.14.0
bitsandbytes>=0.41.0
scipy>=1.10.0
datasets>=2.16.0

# VLM 관련
qwen_vl_utils
torch>=2.4.0
torchvision>=0.19.0

# Flash Attention (별도 설치 필요)
# pip install flash-attn --no-build-isolation

torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
qwen-vl-utils>=0.0.2
pillow>=10.0.0

gradio

opencv-contrib-python

# FastAPI 서버 및 환경 변수 관리
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0